Update to encoded messages instead of extensions
It turns out that extensions are lossy (as in: decoding them in
core loses the information). We could fix this by keeping
track of the original bytes all throughout core, but instead
it's easier to just switch to a byte vector.

This is probably partially why Protobuf 3 switched to the `Any`
type.